feat(aferição): adiciona modificações à migration PL-3570

+++ sql/migrations/progress/migration_adiciona_valores_medida_afericao_PL-3570.sql
@@ -7,8 +7,8 @@ drop function func_afericao_get_afericoes_avulsas_paginada(f_cod_unidades bigint
 create or replace function func_afericao_get_afericoes_avulsas_paginada(f_cod_unidades bigint[],
                                                                         f_data_inicial date,
                                                                         f_data_final date,
-                                                                        f_limit bigint,
-                                                                        f_offset bigint,
+                                                                        f_limit integer,
+                                                                        f_offset integer,
                                                                         f_incluir_medidas boolean default false)
     returns table
             (
@@ -78,8 +78,8 @@ create or replace function func_afericao_get_afericoes_placas_paginada(f_cod_uni
                                                                        f_cod_veiculo bigint,
                                                                        f_data_inicial date,
                                                                        f_data_final date,
-                                                                       f_limit bigint,
-                                                                       f_offset bigint,
+                                                                       f_limit integer,
+                                                                       f_offset integer,
                                                                        f_incluir_medidas boolean default false)
     returns table
             (
